Moves all test models of the Example App Website Pages to their own namespace.
