+++ Source/ExampleApp.Test.Functional/ExampleApp.Test.Functional.csproj
@@ -85,14 +85,14 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Versioning.feature</DependentUpon>
     </Compile>
-    <Compile Include="Models\HomePage.cs" />
+    <Compile Include="Models\ExampleAppPages\HomePage.cs" />
     <Compile Include="Models\SeleniumWebBase.cs" />
-    <Compile Include="Models\SlinqyExampleWebPage.cs" />
-    <Compile Include="Models\SlinqyExampleWebsite.cs" />
+    <Compile Include="Models\ExampleAppPages\SlinqyExampleWebPage.cs" />
+    <Compile Include="Models\ExampleAppPages\SlinqyExampleWebsite.cs" />
     <Compile Include="Models\WebBrowser.cs" />
     <Compile Include="Models\WebPage.cs" />
     <Compile Include="Models\WebPageWidget.cs" />
-    <Compile Include="Models\WebsiteFooter.cs" />
+    <Compile Include="Models\ExampleAppPages\WebsiteFooter.cs" />
     <Compile Include="Steps\NavigationSteps.cs" />
     <Compile Include="Steps\VersionSteps.cs" />
     <Compile Include="Setup.cs" />
