+++ Source/ExampleApp.Test.Functional/Models/ExampleAppPages/SlinqyExampleWebPage.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using OpenQA.Selenium;
 
-namespace ExampleApp.Test.Functional.Models
+namespace ExampleApp.Test.Functional.Models.ExampleAppPages
 {
     public abstract class SlinqyExampleWebPage : WebPage
     {
