+++ Source/ExampleApp.Test.Functional/Steps/NavigationSteps.cs
@@ -1,4 +1,5 @@
 ﻿using ExampleApp.Test.Functional.Models;
+using ExampleApp.Test.Functional.Models.ExampleAppPages;
 using TechTalk.SpecFlow;
 
 namespace ExampleApp.Test.Functional.Steps
